Рефакторинг + postgresql + миграции
