+++ cmd/main.go
@@ -1,0 +1,9 @@
+package main
+
+import "7-steelNote-bot/internal/bot"
+
+func main() {
+
+	bot.StartBot()
+
+}
