+++ compose.yaml
@@ -1,0 +1,8 @@
+services:
+  database:
+    image: postgres:17-alpine3.20
+    container_name: steelnote_postgres
+    env_file:
+      - .env
+    ports:
+      - "5432:5432"
