+++ dbconfig.yml
@@ -1,0 +1,5 @@
+development:
+  dialect: postgres
+  datasource: postgres://steel:note@localhost:5432/steelnotedb?sslmode=disable
+  dir: migrations
+  table: migrations
